Sending country and external_id in Pixel and Conversions API events
Summary:
The key for country is changed from cn to country, because the Automatic Advanced Matching API returns "country". This way, country will be sent in Pixel and Conversions API events.

External id is added as one of the supported match keys, it corresponds to the customer id in Magento database. This value is hashed in Conversions API events(it's recommended) according to the documentation: https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/customer-information-parameters#external-id

ServerSideHelperTest is updated.

Reviewed By: HeyMultiverse

Differential Revision: D24770875

fbshipit-source-id: a83daf1db9aad4515878e3edc2575669ea44b851