Extracting client ip address from other headers and adding validations
Summary:
The key HTTP_X_FORWARDED_FOR in $_SERVER can be a list of ip addresses if the request was handled by a proxy server, so all the processed headers are split by commas.
Each processed ip address must be a valid IPv4 and IPv6 address, not reserved and not private.
More keys in $_SERVER are scanned.
Reference diff: D21049700
Read more: https://itman.in/en/how-to-get-client-ip-address-in-php/

Reviewed By: HeyMultiverse

Differential Revision: D24863723

fbshipit-source-id: 2a2d6b851541cd3fe20a4ae954d34cc73d3d63bc