Adding class to normalize and filter aam fields
Summary:
The class AAMFieldsExtractorHelper is added to filter user data fields based on AAM settings of the pixel and normalize them using the business sdk. If no user data is passed, it returns user data from session.
The method getPixelInitCode() is moved from Common to Head, because it's only used by the Head class.
Most of the test in ServerSideHelper are moved to AAMFieldsExtractorHelper.
It's done based on the discussion on D24902308 .

Reviewed By: marksliva

Differential Revision: D24953889

fbshipit-source-id: 9bb7c63580fc8461201e3a15ebfb3de1c7e61c31