Honor access token passed as an argument for Feed API method
Summary:
Some sellers may want to use our Commerce version for Pixel/Conversions API. We've tested and it works well so far but found one bug with feed upload not working properly because we're not passing access token. This fix is to address it.

Note: it's good to fix this bug but a larger question is if we want to allow product upload within FBE flow in the Commerce version. Sellers could have already set up catalog and feed for Onsite Checkout outside of FBE. Having 2 uploads will cause confusion and possibly errors.

Reviewed By: xiangminxufsu

Differential Revision: D28048656

fbshipit-source-id: e81a943fbc3245f558682fbb456abc5589d34d97