Explicitly request first and last name
Summary:
When the order is imported in Magento, we're splitting the name field on the shipping address by space, using the first bit as the first name and the second bit as the last name. This will not work for cases like this one:  `Mr C A Jennions`.

We need to modify the request to pull first name and last name explicitly:

```
/<ORDER>?fields=shipping_address{first_name, last_name}
```

Response:
```{
},"shipping_address": {"first_name": "Mr","last_name": "C A Jennions "
"id": "352005933368854"
}
```

Reviewed By: amrendra18

Differential Revision: D31945474

fbshipit-source-id: e25556b039656fb9191e2f67421ec046003d830c