Assign variants info to order item
Summary:
Today, an order item for a Magento configurable product (product with variants) is missing info like size and color. This is causing issues for some sellers as their reporting/order fulfillment customizations expect this info to be available with order details.

We need to add info about selected size/color to an order item.

[Asana task](https://app.asana.com/0/1201881326670542/1201867650750508/f)

Reviewed By: dahcsguy

Differential Revision: D34700084

fbshipit-source-id: bc155e727e0260ab02c5dfd3d19be5b3ccd4e6cd