Revamp onboarding mechanism
Summary:
This is an early prototype of scaled onboarding mechanism. Not good for production but OK to use for development and testing.

* Resurfaced the onboarding flow, updated FB login dialog JS code
* Switched to using our 1p app (1843036385894568)
* Previously the access token obtained was a page token. Now we removed the CMS dependency on the page and can use user token instead of page token.

Note: this flow returns a SHORT-LIVED user token, which expires after one hour or so. It's fine for testing but for production we'll need to implement On Behalf API to get a long-lived access token.

Reviewed By: dahcsguy

Differential Revision: D37423887

fbshipit-source-id: e66880a142fbf7407c706f4224bfe4b94cbdb465